# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Post {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  isBreaking: boolean;
  createdAt: string;
  reactionCount: number;
  commentCount: number;
}

export interface PostInput {
  title: string;
  content: string;
  excerpt?: string;
  imageUrl?: string;
  category: string;
  isBreaking?: boolean;
}

export interface PostUpdate {
  title?: string;
  content?: string;
  excerpt?: string;
  imageUrl?: string;
  category?: string;
  isBreaking?: boolean;
}

export type PostStatsByCategoryItem = {
  category: string;
  count: number;
};

export interface PostStats {
  totalPosts: number;
  byCategory: PostStatsByCategoryItem[];
}

export interface BreakingNews {
  text: string;
  enabled: boolean;
}

export interface BreakingNewsInput {
  text: string;
  enabled: boolean;
}

export interface Category {
  id: number;
  name: string;
}

export interface CategoryInput {
  name: string;
}

export interface ReactionCounts {
  like: number;
  laugh: number;
  shock: number;
  angry: number;
}

export type ReactionInputType = typeof ReactionInputType[keyof typeof ReactionInputType];


export const ReactionInputType = {
  like: 'like',
  laugh: 'laugh',
  shock: 'shock',
  angry: 'angry',
} as const;

export interface ReactionInput {
  type: ReactionInputType;
  userId: string;
}

export interface Comment {
  id: number;
  /** @nullable */
  postId?: number | null;
  /** @nullable */
  voteCardId?: number | null;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  parentCommentId?: number | null;
  content: string;
  likeCount: number;
  createdAt: string;
  replies?: Comment[];
}

export interface CommentInput {
  content: string;
  userId: string;
  /** @nullable */
  parentCommentId?: number | null;
}

export type GistStatus = typeof GistStatus[keyof typeof GistStatus];


export const GistStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Gist {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  status: GistStatus;
  createdAt: string;
  /** @nullable */
  publishedAt?: string | null;
  /** @nullable */
  publishAt?: string | null;
}

export interface GistInput {
  content: string;
  imageUrl?: string;
  category: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  commentCount?: number;
  voteCount?: number;
  totalPoints?: number;
  badge?: string;
  createdAt: string;
}

export interface UserInput {
  name: string;
  email: string;
  referredBy?: string;
}

export interface AdminStats {
  totalVisitors: number;
  registeredUsers: number;
  totalPosts: number;
  totalGistsPublished: number;
  totalComments: number;
  pendingGists: number;
  totalVoteCards: number;
  pendingGoatNominees: number;
}

export interface VoteCard {
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageUrl2?: string | null;
  optionALabel: string;
  optionBLabel: string;
  optionACount: number;
  optionBCount: number;
  isActive: boolean;
  createdAt: string;
  totalVotes: number;
  commentCount?: number;
}

export interface VoteCardDetail {
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageUrl2?: string | null;
  optionALabel: string;
  optionBLabel: string;
  optionACount: number;
  optionBCount: number;
  isActive: boolean;
  createdAt: string;
  totalVotes: number;
  commentCount?: number;
  /** @nullable */
  userVote: string | null;
}

export interface VoteCardInput {
  title: string;
  imageUrl?: string;
  imageUrl2?: string;
  optionALabel: string;
  optionBLabel: string;
}

export interface VoteCardUpdate {
  title?: string;
  imageUrl?: string;
  imageUrl2?: string;
  optionALabel?: string;
  optionBLabel?: string;
  isActive?: boolean;
}

export type VoteCardVoteInputChosenOption = typeof VoteCardVoteInputChosenOption[keyof typeof VoteCardVoteInputChosenOption];


export const VoteCardVoteInputChosenOption = {
  a: 'a',
  b: 'b',
} as const;

export interface VoteCardVoteInput {
  userId: string;
  chosenOption: VoteCardVoteInputChosenOption;
}

export type TrendingItemType = typeof TrendingItemType[keyof typeof TrendingItemType];


export const TrendingItemType = {
  post: 'post',
  vote_card: 'vote_card',
} as const;

export interface TrendingItem {
  type: TrendingItemType;
  id: number;
  title: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  category?: string | null;
  engagementScore: number;
  createdAt: string;
  /** @nullable */
  excerpt?: string | null;
}

export interface LeaderboardEntry {
  rank: number;
  id: string;
  name: string;
  totalPoints: number;
  badge: string;
  commentCount: number;
  voteCount: number;
}

export interface GoatCategory {
  id: number;
  name: string;
  createdAt: string;
}

export interface GoatCategoryInput {
  name: string;
}

export interface GoatNominee {
  id: number;
  goatCategoryId: number;
  name: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  description?: string | null;
  voteCount: number;
  status: string;
  createdAt: string;
}

export interface GoatNomineeInput {
  goatCategoryId: number;
  name: string;
  photoUrl?: string;
  description?: string;
  submittedBy?: string;
}

export interface GoatVoteInput {
  userId: string;
  nomineeId: number;
}

export interface GoatCategoryWithNominees {
  id: number;
  name: string;
  nominees: GoatNominee[];
  /** @nullable */
  userVotedNomineeId: number | null;
}

export type ListPostsParams = {
category?: string;
limit?: number;
offset?: number;
};

export type ListPublicGistsParams = {
category?: string;
limit?: number;
offset?: number;
};

export type ListVoteCardsParams = {
/**
 * If true, include inactive cards (admin)
 */
all?: boolean;
};

export type GetVoteCardParams = {
userId?: string;
};

export type GetTrendingParams = {
limit?: number;
};

export type ListGoatNomineesParams = {
userId?: string;
};