# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vawulence City Media API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a post (sorted by likes desc)
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
...
  /gists:
    get:
      operationId: listPublicGists
      tags: [gists]
      summary: List approved/published gists
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gist"
    post:
      operationId: submitGist
      tags: [gists]
      summary: Submit a gist (anonymous, goes to pending queue)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GistInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gist"
...
  /gists/{id}/approve:
    post:
      operationId: approveGist
      tags: [gists]
      summary: Approve and publish a gist (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gist"
  /gists/{id}/reject:
    post:
      operationId: rejectGist
      tags: [gists]
      summary: Reject a gist (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gist"

  # ── Users ─────────────────────────────────────────────────────────────────────
...
  /goat/categories:
    get:
      operationId: listGoatCategories
      tags: [goat]
      summary: List GOAT categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoatCategory"
    post:
      operationId: createGoatCategory
      tags: [goat]
      summary: Create a GOAT category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoatCategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatCategory"
  /goat/categories/{id}:
    delete:
      operationId: deleteGoatCategory
      tags: [goat]
      summary: Delete a GOAT category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /goat/categories/{id}/nominees:
    get:
      operationId: listGoatNominees
      tags: [goat]
      summary: List approved nominees for a GOAT category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatCategoryWithNominees"
  /goat/categories/{id}/vote:
    post:
      operationId: voteGoat
      tags: [goat]
      summary: Vote for a nominee in a GOAT category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoatVoteInput"
      responses:
        "200":
          description: Updated nominees list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatCategoryWithNominees"
        "409":
          description: Already voted in this category
  /goat/nominees:
    post:
      operationId: submitGoatNominee
      tags: [goat]
      summary: Submit a nominee for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoatNomineeInput"
      responses:
        "201":
          description: Submitted (pending review)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatNominee"
...
  /goat/nominees/{id}/approve:
    post:
      operationId: approveGoatNominee
      tags: [goat]
      summary: Approve a GOAT nominee (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatNominee"
  /goat/nominees/{id}/reject:
    post:
      operationId: rejectGoatNominee
      tags: [goat]
      summary: Reject a GOAT nominee (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GoatNominee"
components:
  schemas:
...
    Comment:
      type: object
      required: [id, content, likeCount, createdAt]
      properties:
        id:              { type: integer }
        postId:          { type: ["integer", "null"] }
        voteCardId:      { type: ["integer", "null"] }
        userId:          { type: ["string", "null"] }
        userName:        { type: ["string", "null"] }
        parentCommentId: { type: ["integer", "null"] }
        content:         { type: string }
        likeCount:       { type: integer }
        createdAt:       { type: string }
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
    CommentInput:
      type: object
      required: [content, userId]
      properties:
        content:         { type: string }
        userId:          { type: string }
        parentCommentId: { type: ["integer", "null"] }

    # ── Gist ──────────────────────────────────────────────────────────────────
    Gist:
      type: object
      required: [id, content, category, status, createdAt]
      properties:
        id:          { type: integer }
        content:     { type: string }
        imageUrl:    { type: ["string", "null"] }
        category:    { type: string }
        status:      { type: string, enum: [pending, approved, rejected] }
        createdAt:   { type: string }
        publishedAt: { type: ["string", "null"] }
        publishAt:   { type: ["string", "null"] }
    GistInput:
      type: object
      required: [content, category]
      properties:
        content:  { type: string }
        imageUrl: { type: string }
        category: { type: string }

    # ── User ──────────────────────────────────────────────────────────────────
...
    GoatCategory:
      type: object
      required: [id, name, createdAt]
      properties:
        id:        { type: integer }
        name:      { type: string }
        createdAt: { type: string }
    GoatCategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
    GoatNominee:
      type: object
      required: [id, goatCategoryId, name, voteCount, status, createdAt]
      properties:
        id:             { type: integer }
        goatCategoryId: { type: integer }
        name:           { type: string }
        photoUrl:       { type: ["string", "null"] }
        description:    { type: ["string", "null"] }
        voteCount:      { type: integer }
        status:         { type: string }
        createdAt:      { type: string }
    GoatNomineeInput:
      type: object
      required: [goatCategoryId, name]
      properties:
        goatCategoryId: { type: integer }
        name:           { type: string }
        photoUrl:       { type: string }
        description:    { type: string }
        submittedBy:    { type: string }
    GoatVoteInput:
      type: object
      required: [userId, nomineeId]
      properties:
        userId:    { type: string }
        nomineeId: { type: integer }
    GoatCategoryWithNominees:
      type: object
      required: [id, name, nominees, userVotedNomineeId]
      properties:
        id:                 { type: integer }
        name:               { type: string }
        nominees:
          type: array
          items:
            $ref: "#/components/schemas/GoatNominee"
        userVotedNomineeId: { type: ["integer", "null"] }