# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  /** @nullable */
  postId?: number | null;
  /** @nullable */
  voteCardId?: number | null;
  /** @nullable */
  userId?: string | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  parentCommentId?: number | null;
  content: string;
  likeCount: number;
  createdAt: string;
  replies?: Comment[];
}