# lib/api-zod/src/generated/types/goatCategoryWithNominees.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */
import type { GoatNominee } from './goatNominee';

export interface GoatCategoryWithNominees {
  id: number;
  name: string;
  nominees: GoatNominee[];
  /** @nullable */
  userVotedNomineeId: number | null;
}