# lib/api-zod/src/generated/types/goatVoteInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vawulence City Media API
 * OpenAPI spec version: 0.1.0
 */

export interface GoatVoteInput {
  userId: string;
  nomineeId: number;
}